+++ src/formatters/types.ts
@@ -2,6 +2,7 @@ export type FormatterName = 'rubyfmt' | 'rufo';
 
 export interface FormatterDescriptor {
   readonly id: FormatterName;
+  readonly injectsTrailingNewline: boolean;
   readonly installUrl: string;
   readonly versionArgs?: string[];
 }